Converted FlightAware's flight logs to GPX files
Created a Python script to automate the conversion of the HTML table to a GPX track. Formatted with Black, but hasn't been linted.

Another Python script can find (via brute-force) the URL of a certain flight in a certain day. Useful if the logs for a specific flight weren't recorded.

Tracks that cover multiple countries are added to a separate layer.
FUTURE: The "international" layer should be disabled by default.