Added a custom Control for changing the LineString colour style
The new Control is made up of a div containing a title, drop-down select box, and legend content. It provides helper functions for modifying the content.

To allow the colour style to change, the GeoJSON content must be re-parsed with the new styles/thresholds. This also means that the map content must be cleared and re-drawn.

Significant refactoring was needed to allow re-drawing the map content. Partial functions are used in some places so avoid passing long parameter lists to functions.

Other improvements:
* All Controls are accessible from the Map object.
* The list of LineString styles is customisable.
* Errors during the fetch or parsing of the GeoJSON will show an alert.