Attempted to extend the GPS Visualizer functions to show the GeoJSON data
Giving up on this approach. the GPSVisualizer functions seem to be very hard to extend (or even understand in some cases). The data seems to require specific formats, which would add extra processing. There's also lots of code that wouldn't be needed in the end.

Analysed the tracklist functions, and these don't seem too hard to replicate (famous last words). It'd be nice to turn the library into a JavaScript module, so that the new Leaflet Controls would be in separate files.

This commit also downloaded the GPS Visualizer function file.