Converted the original Leaflet POC to a JavaScript module
The module is split into multiple files by functionality. Each Leaflet Control is defined in its own file, and each Layer type is created in its own file.

The main entry-point remains in "map_leaflet.js". The file hasn't been renamed in this commit so that Git sees a file modification instead of a deletion/creation.

The client configuration (DOM ID for the map, GeoJSON path, and lineString styles) are now defined directly by clients. In this case, in the HTML file.