Split CSS-in-JS to separate SCSS files
This allows the styles to be more easily understood, since they're all defined in one place for a component.
Applying styles via classes instead of per-element is also a more traditional way of styling elements.

The Rollup command compiles the Sass to a single CSS file.
Each Sass file must be imported from the TypeScript source.
Note that importing the Sass does not allow the class names to be used within TypeScript.